% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvfunOnFun.R
\name{cvfunOnFun}
\alias{cvfunOnFun}
\title{A function to choose optimal number of eigenfunctions via Cross-Validation}
\usage{
cvfunOnFun(seed, predictor, response, actual, folds = 10)
}
\arguments{
\item{seed}{Seed for reproducibility}

\item{predictor}{An object of class irregMFPCA corresponding to the predictor}

\item{response}{An object of class irregMFPCA corresponding to the response}

\item{actual}{A data frame with the actual values of the response}

\item{folds}{Number of folds for cross-validation. Default is 10}
}
\value{
A matrix of mean squared errors for each combination of number of eigenfunctions
}
\description{
A function to choose optimal number of eigenfunctions via Cross-Validation
}
