% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvfunOnFun.R
\name{.closest_t}
\alias{.closest_t}
\title{Internal function for matching to closest time point}
\usage{
.closest_t(actual_time, hat_time)
}
\arguments{
\item{actualhat}{Time points of estimated functions}

\item{actual}{Time points of actual functions}
}
\value{
A vector of indices of the closest time points
}
\description{
Internal function for matching to closest time point
}
