% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpcaFormat.R
\name{.create_tibble}
\alias{.create_tibble}
\title{A function that wraps variables into a tibble}
\usage{
.create_tibble(...)
}
\arguments{
\item{...}{A list of vectors of variables}
}
\value{
A tibble of variables
}
\description{
A function that wraps variables into a tibble
}
