% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMFPCA.R
\name{.generate_mu}
\alias{.generate_mu}
\title{A function to generate mean functions for each variable}
\usage{
.generate_mu(t, func_list)
}
\arguments{
\item{t}{A vector of time points}

\item{func_list}{A list of functions to generate mean functions}
}
\value{
A t by num.var matrix of mean functions
}
\description{
A function to generate mean functions for each variable
}
