% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMFPCA.R
\name{.generate_phi}
\alias{.generate_phi}
\title{A function to generate orthogonal eigenfunctions}
\usage{
.generate_phi(t, eigen_funs)
}
\arguments{
\item{t}{A vector of time points}

\item{eigen_funs}{A list of functions to generate eigenfunctions; Length should match number of components}
}
\value{
A t by number of components matrix of eigenfunctions such that the inner product is length of t * \eqn{I_n}
}
\description{
A function to generate orthogonal eigenfunctions
}
