% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMFPCA.R
\name{.generate_xi}
\alias{.generate_xi}
\title{A function to generate orthonormal eigenscores}
\usage{
.generate_xi(n, components)
}
\arguments{
\item{n}{Number of subjects}

\item{components}{Number of components}
}
\value{
A n by length(t) matrix of orthogonal eigenscores such that the inner product is \eqn{n \times I_n}
}
\description{
A function to generate orthonormal eigenscores
}
