% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibbleFormat.R
\name{.longFormat}
\alias{.longFormat}
\title{A function to convert a data frame into a long format}
\usage{
.longFormat(dat)
}
\arguments{
\item{dat}{A data frame of data}
}
\value{
A data frame in long format
}
\description{
A function to convert a data frame into a long format
}
