% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMFPCA.R
\name{.normalizer}
\alias{.normalizer}
\title{A function to normalize a matrix}
\usage{
.normalizer(mat, normal = NULL)
}
\arguments{
\item{mat}{A matrix}

\item{normal}{A scalar to normalize the matrix. Default is the number of rows of the matrix}
}
\value{
A matrix normalized so that its inner product is \eqn{normal*I_n}
}
\description{
A function to normalize a matrix
}
