% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMFPCA.R
\name{.stack_phi}
\alias{.stack_phi}
\title{A function to stack multiple eigenfunctions if there are multiple variables}
\usage{
.stack_phi(t, eigen_funs_list)
}
\arguments{
\item{t}{A vector of time points}

\item{eigen_funs_list}{A list of lists of functions to generate eigenfunctions for each variable}
}
\value{
A stacked matrix of eigenfunctions
}
\description{
A function to stack multiple eigenfunctions if there are multiple variables
}
