% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpcaFormat.R
\name{.vector_filler}
\alias{.vector_filler}
\title{A function to fill an empty vector with values from a data frame}
\usage{
.vector_filler(v, target, source, uid)
}
\arguments{
\item{v}{A vector to fill}

\item{target}{The target column to fill}

\item{source}{The source data frame}

\item{uid}{A unique identifier for each subject}
}
\value{
A filled vector
}
\description{
A function to fill an empty vector with values from a data frame
}
