% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funOnFun.R
\name{funOnFun}
\alias{funOnFun}
\title{Function to conduct Function on Function regression based on eigenscores}
\usage{
funOnFun(resp, pred)
}
\arguments{
\item{resp}{A response `irregMFPCA` object}

\item{pred}{A predictor `irregMFPCA` object}
}
\value{
The regression matrix
}
\description{
Function to conduct Function on Function regression based on eigenscores
}
