% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irregMFPCA.R
\name{irregMFPCA}
\alias{irregMFPCA}
\title{A function to perform irregular MFPCA}
\usage{
irregMFPCA(components = NULL, split = F, ...)
}
\arguments{
\item{components}{Number of principal components to consider}

\item{split}{Whether to split the data into separate variables}

\item{...}{A list of FPCA objects}
}
\value{
A list containing the unstacked (wide) eigenfunctions, stacked (long) eigenfunctions, eigenscores, Dhat, and Xhat
}
\description{
A function to perform irregular MFPCA
}
