% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mip.R
\name{mip}
\alias{mip}
\title{Compute Inner Product (Covariance) of Matrices}
\usage{
mip(mat)
}
\arguments{
\item{mat}{An n x p matrix}
}
\value{
A p x p matrix of the inner product of the columns of mat
}
\description{
Compute Inner Product (Covariance) of Matrices
}
