% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconBeta.R
\name{populationBeta}
\alias{populationBeta}
\title{A function to reconstruct the estimand, the population beta matrix, from the predictor `simMFPCA` object}
\usage{
populationBeta(popCovY, popB, t, X)
}
\arguments{
\item{popCovY}{The population covariance matrix of the response}

\item{popB}{The population B matrix}

\item{t}{The time points}

\item{X}{The predictor `simMFPCA` object}
}
\value{
The reconstructed population beta matrix
}
\description{
A function to reconstruct the estimand, the population beta matrix, from the predictor `simMFPCA` object
}
