% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMFPCA.R
\name{populationEigen}
\alias{populationEigen}
\title{Return Orthogonalized Population Quantities for Non-orthogonal Data Rotated by a matrix B}
\usage{
populationEigen(Y, B, t, sigma)
}
\arguments{
\item{Y}{A simMFPCA object}

\item{B}{A matrix to rotate the data}

\item{t}{A vector of time points}

\item{sigma}{A scalar for the noise}
}
\value{
A list of (orthogonalized) population covariance, eigenfunctions, eigenscores, and B
}
\description{
Return Orthogonalized Population Quantities for Non-orthogonal Data Rotated by a matrix B
}
