% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funOnFun.R
\name{populationReg}
\alias{populationReg}
\title{Function to return estimand of `funOnFun`}
\usage{
populationReg(X, popXiY)
}
\arguments{
\item{X}{A predictor `simMFPCA` object}

\item{popXiY}{The population orthogonalized eigenscores of the response}
}
\value{
The estimand of the regression matrix
}
\description{
Function to return estimand of `funOnFun`
}
