% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconBeta.R
\name{reconBeta}
\alias{reconBeta}
\title{A function to reconstruct the beta matrix from the response and predictor `irregMFPCA` objects}
\usage{
reconBeta(resp, pred, Bhat)
}
\arguments{
\item{resp}{A response `irregMFPCA` object}

\item{pred}{A predictor `irregMFPCA` object}

\item{Bhat}{The estimated B matrix from the `funOnFun` function}
}
\value{
The reconstructed estimated beta matrix
}
\description{
A function to reconstruct the beta matrix from the response and predictor `irregMFPCA` objects
}
