% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularizeTime.R
\name{regularizeTime}
\alias{regularizeTime}
\title{A function to regularize time for each id in a dataset}
\usage{
regularizeTime(dat, id_col)
}
\arguments{
\item{dat}{A data frame with clustered data}

\item{id_col}{The name of the column with the id}
}
\value{
A data frame with a new column t that goes from 0 to 1 for each id
}
\description{
A function to regularize time for each id in a dataset
}
