% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMiss.R
\name{simMiss}
\alias{simMiss}
\title{A function to simulate missing data}
\usage{
simMiss(n_obs, dat, t = seq(0, 1, length.out = 51), seed = 51)
}
\arguments{
\item{n_obs}{Number of observations to keep}

\item{dat}{A matrix of data}

\item{t}{A vector of time points}

\item{seed}{A seed for random number generation}
}
\value{
A matrix of data with missing values
}
\description{
A function to simulate missing data
}
