% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibbleFormat.R
\name{tibbleFormat}
\alias{tibbleFormat}
\title{A function to convert a matrix into a tibble}
\usage{
tibbleFormat(dat, t)
}
\arguments{
\item{dat}{A matrix of data}

\item{t}{A vector of time points}
}
\value{
A tibble of data
}
\description{
A function to convert a matrix into a tibble
}
