% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpcaFormat.R
\name{fpcaFormat}
\alias{fpcaFormat}
\title{A function to format data for FPCA}
\usage{
fpcaFormat(dat)
}
\arguments{
\item{dat}{A data frame of data}
}
\value{
A tibble of data in FPCA format
}
\description{
A function to format data for FPCA
}
